pyproject.toml
[project]
name = "curly-octo-guacamole"
version = "0.1.0"
description = "Testing Accelerators for fptmark"
authors = [
    {name = "Nick Baynham", email = "nickbaynham@gmail.com"},
]
dependencies = ["requests>=2.32.4", "beautifulsoup4>=4.13.4", "pytest-playwright>=0.7.0", "python-dotenv>=1.0.0", "dotenv>=0.9.9", "pymongo>=4.13.2", "pytest-asyncio>=1.1.0"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "NONE"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short -s"
console_output_style = "auto"
